moved cleanUpOnDisconnect to CoreActivity...it wasn't getting called if the network fragment wasn't open introduced a new HOST_DISBANDED intent action to clean up after a host is disbanded
added test for null to searchView in onCreateOptionsMenu in CoreActivity..this can probably be removed once the SearchView stuff is sorted out.
